Move api/server/httputils to daemon/server/httputils
