+++ api/server/middleware.go
@@ -2,7 +2,7 @@ package server
 
 import (
 	"github.com/containerd/log"
-	"github.com/docker/docker/api/server/httputils"
+	"github.com/docker/docker/daemon/server/httputils"
 	"github.com/docker/docker/daemon/server/middleware"
 )
