+++ daemon/server/httputils/write_log_stream.go
