+++ daemon/server/router/checkpoint/checkpoint.go
@@ -1,7 +1,7 @@
 package checkpoint
 
 import (
-	"github.com/docker/docker/api/server/httputils"
+	"github.com/docker/docker/daemon/server/httputils"
 	"github.com/docker/docker/daemon/server/router"
 )
