+++ daemon/server/router/container/container.go
@@ -1,7 +1,7 @@
 package container
 
 import (
-	"github.com/docker/docker/api/server/httputils"
+	"github.com/docker/docker/daemon/server/httputils"
 	"github.com/docker/docker/daemon/server/router"
 )
