+++ daemon/server/router/debug/debug.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"net/http/pprof"
 
-	"github.com/docker/docker/api/server/httputils"
+	"github.com/docker/docker/daemon/server/httputils"
 	"github.com/docker/docker/daemon/server/router"
 )
