+++ daemon/server/router/distribution/distribution_routes.go
@@ -10,8 +10,8 @@ import (
 	"github.com/docker/distribution/manifest/manifestlist"
 	"github.com/docker/distribution/manifest/schema1"
 	"github.com/docker/distribution/manifest/schema2"
-	"github.com/docker/docker/api/server/httputils"
 	"github.com/docker/docker/api/types/registry"
+	"github.com/docker/docker/daemon/server/httputils"
 	distributionpkg "github.com/docker/docker/distribution"
 	"github.com/docker/docker/errdefs"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
