+++ daemon/server/router/swarm/helpers.go
@@ -5,12 +5,12 @@ import (
 	"errors"
 	"net/http"
 
-	"github.com/docker/docker/api/server/httputils"
 	basictypes "github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/backend"
 	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/swarm"
 	"github.com/docker/docker/api/types/versions"
+	"github.com/docker/docker/daemon/server/httputils"
 )
 
 // swarmLogs takes an http response, request, and selector, and writes the logs
